refactor: 예외발생 리팩토링, REST API 엔드포인트 개선 (GS-66)
임시로 만들어둔 예외발생 클래스를 리팩토링했습니다. 엔드포인트의 의도가 더 잘 드러날 수 있도록 개선했습니다.
