+++ src/main/java/com/foo/gosucatcher/domain/member/application/MemberProfileService.java
@@ -28,13 +28,6 @@ public class MemberProfileService {
 	private final MemberRepository memberRepository;
 	private final PasswordEncoder passwordEncoder;
 
-	public void deleteMember(Long memberId) {
-		Member member = memberRepository.findById(memberId)
-			.orElseThrow(() -> new EntityNotFoundException(ErrorCode.NOT_FOUND_MEMBER));
-
-		memberRepository.delete(member);
-	}
-
 	public MemberProfileResponse findMemberProfile(Long memberId) {
 		Member member = memberRepository.findById(memberId)
 			.orElseThrow(() -> new EntityNotFoundException(ErrorCode.NOT_FOUND_MEMBER));
