+++ src/main/java/com/foo/gosucatcher/global/error/ErrorCode.java
@@ -41,6 +41,7 @@ public enum ErrorCode {
 	UNAUTHENTICATED_MEMBER("M006", "인증되지 않은 회원입니다."),
 	INCORRECT_AUTH_NUMBER("M007", "잘못된 인증번호입니다."),
 	EXPIRED_AUTH_NUMBER("M008", "인증시간이 만료되었습니다."),
+	NOT_CREATION_AUTH_MESSAGE("M009", "인증 메시지를 생성할 수 없습니다."),
 
 	//파일
 	NOT_FOUND_IMAGE("F001", "존재하지 않는 이미지 입니다."),
